requirements.txt
"""
Requirements para Revisão de Literatura Open Science
====================================================

Instalar com: pip install -r requirements.txt
"""

# Data Processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.10.10

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.9.0
pdf2image==1.16.3
pytesseract==0.3.10

# LLM APIs - A/B Testing
openai==0.27.8
anthropic==0.3.0
google-generativeai==0.3.0  # NOVO: Gemini API para A/B Testing
langchain==0.0.239

# Data Analysis
scipy==1.11.1
scikit-learn==1.3.0
statsmodels==0.14.0
krippendorff==0.6.1  # NOVO: Krippendorff's Alpha para PRISMA 2024-IA

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
networkx==3.1

# Documentation
markdown==3.4.1
python-frontmatter==1.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.4.1

# A/B Testing & Validação inter-IA
rapidfuzz==3.1.1  # Fuzzy string matching para concordância

# Git & Version Control
GitPython==3.1.32

# Jupyter (Optional, para análises interativas)
jupyter==1.0.0
jupyterlab==4.0.4
